split up tests into view and logic
