+++ web/html/src/manager/admin/setup/products/products.test.tsx
@@ -1,10 +1,7 @@
-import { createElement } from "react";
-
 import { screen } from "@testing-library/react";
 
 import { render } from "utils/test-utils";
 
-import { type ProductLike, getProductSelectionState } from "./product-check/product-selection.utils";
 import { CheckListItem } from "./products";
 import { searchCriteriaInExtension } from "./products.utils";
 
@@ -93,167 +90,6 @@ describe("Testing searchCriteriaInExtension", () => {
   });
 });
 
-describe("getProductSelectionState", () => {
-  const productTree: ProductLike = {
-    identifier: "base",
-    status: "AVAILABLE",
-    extensions: [
-      {
-        identifier: "recommended-child",
-        status: "AVAILABLE",
-        extensions: [
-          {
-            identifier: "grandchild",
-            status: "AVAILABLE",
-            extensions: [],
-          },
-        ],
-      },
-      {
-        identifier: "optional-child",
-        status: "AVAILABLE",
-        extensions: [],
-      },
-    ],
-  };
-
-  test("returns partially when only recommended children are auto-selected", () => {
-    const selectedItems = [{ identifier: "base" }, { identifier: "recommended-child" }];
-
-    expect(getProductSelectionState(productTree, selectedItems)).toBe("partially");
-  });
-
-  test("returns checked when the whole subtree is selected", () => {
-    const selectedItems = [
-      { identifier: "base" },
-      { identifier: "recommended-child" },
-      { identifier: "grandchild" },
-      { identifier: "optional-child" },
-    ];
-
-    expect(getProductSelectionState(productTree, selectedItems)).toBe("checked");
-  });
-
-  test("returns unchecked when neither the product nor its children are selected", () => {
-    expect(getProductSelectionState(productTree, [])).toBe("unchecked");
-  });
-
-  test("returns checked for a child when all of its direct children are selected", () => {
-    const selectedItems = [{ identifier: "recommended-child" }, { identifier: "grandchild" }];
-
-    expect(getProductSelectionState(productTree.extensions![0], selectedItems)).toBe("checked");
-  });
-
-  test("keeps the parent partially selected when a direct child is only partially selected", () => {
-    const selectedItems = [
-      { identifier: "base" },
-      { identifier: "recommended-child" },
-      { identifier: "optional-child" },
-    ];
-
-    expect(getProductSelectionState(productTree, selectedItems)).toBe("partially");
-  });
-
-  test("keeps a selected non-root parent partially selected when not all of its own children are selected", () => {
-    const selectedItems = [{ identifier: "recommended-child" }];
-
-    expect(getProductSelectionState(productTree.extensions![0], selectedItems)).toBe("partially");
-  });
-
-  test("returns partially when a direct child is selected but the parent is not", () => {
-    const selectedItems = [{ identifier: "optional-child" }];
-
-    expect(getProductSelectionState(productTree, selectedItems)).toBe("partially");
-  });
-
-  test("treats installed direct children as selected", () => {
-    const installedTree: ProductLike = {
-      identifier: "installed-parent",
-      status: "AVAILABLE",
-      extensions: [
-        {
-          identifier: "installed-child",
-          status: "INSTALLED",
-          extensions: [],
-        },
-      ],
-    };
-
-    const selectedItems = [{ identifier: "installed-parent" }];
-
-    expect(getProductSelectionState(installedTree, selectedItems)).toBe("checked");
-  });
-
-  test("ignores unavailable direct children when evaluating selection state", () => {
-    const unavailableTree: ProductLike = {
-      identifier: "unavailable-parent",
-      status: "AVAILABLE",
-      extensions: [
-        {
-          identifier: "available-child",
-          status: "AVAILABLE",
-          extensions: [],
-        },
-        {
-          identifier: "unavailable-child",
-          status: "UNAVAILABLE",
-          extensions: [],
-        },
-      ],
-    };
-
-    const selectedItems = [{ identifier: "unavailable-parent" }, { identifier: "available-child" }];
-
-    expect(getProductSelectionState(unavailableTree, selectedItems)).toBe("checked");
-  });
-
-  test("matches the rendered list behavior for a multi-level recommended products tree", () => {
-    const renderedTree: ProductLike = {
-      identifier: "sle-hpc-15-sp2",
-      status: "AVAILABLE",
-      extensions: [
-        {
-          identifier: "ltss-15-sp2",
-          status: "AVAILABLE",
-          extensions: [],
-        },
-        {
-          identifier: "basesystem-15-sp2",
-          status: "AVAILABLE",
-          extensions: [
-            {
-              identifier: "python2-15-sp2",
-              status: "AVAILABLE",
-              extensions: [],
-            },
-            {
-              identifier: "desktop-applications-15-sp2",
-              status: "AVAILABLE",
-              extensions: [],
-            },
-            {
-              identifier: "server-applications-15-sp2",
-              status: "AVAILABLE",
-              extensions: [],
-            },
-          ],
-        },
-      ],
-    };
-
-    const selectedItems: ProductLike[] = [
-      { identifier: "sle-hpc-15-sp2" },
-      { identifier: "ltss-15-sp2" },
-      { identifier: "basesystem-15-sp2" },
-      { identifier: "python2-15-sp2" },
-      { identifier: "desktop-applications-15-sp2" },
-    ];
-
-    expect(getProductSelectionState(renderedTree, selectedItems)).toBe("partially");
-    expect(getProductSelectionState(renderedTree.extensions?.[1] as ProductLike, selectedItems)).toBe("partially");
-  });
-});
-
 describe("CheckListItem checkbox rendering", () => {
   const cols = {
     selector: { width: 2, um: "em" },
@@ -322,20 +158,20 @@ describe("CheckListItem checkbox rendering", () => {
 
   test("renders an indeterminate checkbox when a direct child is only partially selected", () => {
     render(
-      createElement(CheckListItem, {
-        item,
-        bypassProps: buildBypassProps([
+      <CheckListItem
+        item={item}
+        bypassProps={buildBypassProps([
           { identifier: "base" },
           { identifier: "basesystem" },
           { identifier: "python2" },
           { identifier: "desktop-applications" },
-        ]),
-        handleSelectedItems: jest.fn(),
-        handleUnselectedItems: jest.fn(),
-        treeLevel: 1,
-        childrenDisabled: false,
-        index: 0,
-      })
+        ])}
+        handleSelectedItems={jest.fn()}
+        handleUnselectedItems={jest.fn()}
+        treeLevel={1}
+        childrenDisabled={false}
+        index={0}
+      />
     );
 
     const checkbox = screen.getByRole("checkbox") as HTMLInputElement;
@@ -345,21 +181,21 @@ describe("CheckListItem checkbox rendering", () => {
 
   test("renders a checked checkbox when the whole visible subtree is selected", () => {
     render(
-      createElement(CheckListItem, {
-        item,
-        bypassProps: buildBypassProps([
+      <CheckListItem
+        item={item}
+        bypassProps={buildBypassProps([
           { identifier: "base" },
           { identifier: "basesystem" },
           { identifier: "python2" },
           { identifier: "desktop-applications" },
           { identifier: "workstation-extension" },
-        ]),
-        handleSelectedItems: jest.fn(),
-        handleUnselectedItems: jest.fn(),
-        treeLevel: 1,
-        childrenDisabled: false,
-        index: 0,
-      })
+        ])}
+        handleSelectedItems={jest.fn()}
+        handleUnselectedItems={jest.fn()}
+        treeLevel={1}
+        childrenDisabled={false}
+        index={0}
+      />
     );
 
     const checkbox = screen.getByRole("checkbox") as HTMLInputElement;
@@ -369,20 +205,20 @@ describe("CheckListItem checkbox rendering", () => {
 
   test("renders an installed product as a checked disabled checkbox", () => {
     render(
-      createElement(CheckListItem, {
-        item: {
+      <CheckListItem
+        item={{
           ...item,
           identifier: "installed-base",
           status: "INSTALLED",
           extensions: [],
-        },
-        bypassProps: buildBypassProps([]),
-        handleSelectedItems: jest.fn(),
-        handleUnselectedItems: jest.fn(),
-        treeLevel: 1,
-        childrenDisabled: false,
-        index: 0,
-      })
+        }}
+        bypassProps={buildBypassProps([])}
+        handleSelectedItems={jest.fn()}
+        handleUnselectedItems={jest.fn()}
+        treeLevel={1}
+        childrenDisabled={false}
+        index={0}
+      />
     );
 
     const checkbox = screen.getByRole("checkbox") as HTMLInputElement;
